+++ Engine/Source/Runtime/Core/Private/Math/UnrealMath.cpp
@@ -1791,8 +1791,8 @@ static bool ComputeProjectedSphereShaft(
 	int32& InOutMaxX
 	)
 {
-	float ViewX = InOutMinX;
-	float ViewSizeX = InOutMaxX - InOutMinX;
+	float ViewX = (float)InOutMinX;
+	float ViewSizeX = (float)(InOutMaxX - InOutMinX);
 
 	// Vertical planes: T = <Nx, 0, Nz, 0>
 	float Discriminant = (FMath::Square(LightX) - FMath::Square(Radius) + FMath::Square(LightZ)) * FMath::Square(LightZ);
@@ -1816,11 +1816,11 @@ static bool ComputeProjectedSphereShaft(
 			float X = (Dot3(P,Axis) / P.W + 1.0f * AxisSign) / 2.0f * AxisSign;
 			if(FMath::IsNegativeFloat(Nxa) ^ FMath::IsNegativeFloat(AxisSign))
 			{
-				InOutMaxX = FMath::Min<int64>(FMath::CeilToInt(ViewSizeX * X + ViewX),InOutMaxX);
+				InOutMaxX = FMath::Min<int32>(FMath::CeilToInt(ViewSizeX * X + ViewX),InOutMaxX);
 			}
 			else
 			{
-				InOutMinX = FMath::Max<int64>(FMath::FloorToInt(ViewSizeX * X + ViewX),InOutMinX);
+				InOutMinX = FMath::Max<int32>(FMath::FloorToInt(ViewSizeX * X + ViewX),InOutMinX);
 			}
 		}
 
@@ -1832,11 +1832,11 @@ static bool ComputeProjectedSphereShaft(
 			float X = (Dot3(P,Axis) / P.W + 1.0f * AxisSign) / 2.0f * AxisSign;
 			if(FMath::IsNegativeFloat(Nxb) ^ FMath::IsNegativeFloat(AxisSign))
 			{
-				InOutMaxX = FMath::Min<int64>(FMath::CeilToInt(ViewSizeX * X + ViewX),InOutMaxX);
+				InOutMaxX = FMath::Min<int32>(FMath::CeilToInt(ViewSizeX * X + ViewX),InOutMaxX);
 			}
 			else
 			{
-				InOutMinX = FMath::Max<int64>(FMath::FloorToInt(ViewSizeX * X + ViewX),InOutMinX);
+				InOutMinX = FMath::Max<int32>(FMath::FloorToInt(ViewSizeX * X + ViewX),InOutMinX);
 			}
 		}
 	}
@@ -2556,9 +2556,9 @@ CORE_API FQuat FMath::QInterpTo(const FQuat& Current, const FQuat& Target, float
 
 CORE_API float ClampFloatTangent( float PrevPointVal, float PrevTime, float CurPointVal, float CurTime, float NextPointVal, float NextTime )
 {
-	const float PrevToNextTimeDiff = FMath::Max< double >( KINDA_SMALL_NUMBER, NextTime - PrevTime );
-	const float PrevToCurTimeDiff = FMath::Max< double >( KINDA_SMALL_NUMBER, CurTime - PrevTime );
-	const float CurToNextTimeDiff = FMath::Max< double >( KINDA_SMALL_NUMBER, NextTime - CurTime );
+	const float PrevToNextTimeDiff = FMath::Max< float >( KINDA_SMALL_NUMBER, NextTime - PrevTime );
+	const float PrevToCurTimeDiff = FMath::Max< float >( KINDA_SMALL_NUMBER, CurTime - PrevTime );
+	const float CurToNextTimeDiff = FMath::Max< float >( KINDA_SMALL_NUMBER, NextTime - CurTime );
 
 	float OutTangentVal = 0.0f;
 
@@ -3238,7 +3238,7 @@ PrecLoop:
 				}
 				else
 				{
-					V = (int32)V % (int32)W;
+					V = (float)((int32)V % (int32)W);
 					c = GrabChar(pStr);
 					goto PrecLoop;
 				}
@@ -3487,7 +3487,7 @@ namespace FMathPerlinHelpers
 	// Curve w/ second derivative vanishing at 0 and 1, from Perlin's improved noise paper
 	FORCEINLINE float SmoothCurve(float X)
 	{
-		return X * X * X * (X * (X * 6.0 - 15.0) + 10.0);
+		return X * X * X * (X * (X * 6.0f - 15.0f) + 10.0f);
 	}
 };
